feat: Initial commit of meh
This commit is the first one for meh, or Michael's Error Handling. It's
currently an experimental proof of concept library to test out how we
could do error handling and control flow in Rust if features like try
blocks and the Try trait become stabilized. Until then it will stay in a
pre-1.0 state and will require a nightly compiler.

For the initial bit of code we implement a `Retryable` type. Much like
Result it has an Ok and an Err variant, but it also has a third one
Retry. With this users can use ? such that it will pass up errors that
should be passed up the stack or return a `Retry` type with just the
`Ok` and `Retry` variants so that control flow can be done to retry
right then and there. The Retry type does not implement Try and so it's
incapable of being passed up the stack easily and is a forcing function
to handle the retryable error if it does happen.